+++ src/types/popover.ts
@@ -1,0 +1,14 @@
+export const PopoverPlacementVariants = [
+  'top',
+  'right',
+  'bottom',
+  'left',
+  'top-start',
+  'top-end',
+  'right-start',
+  'right-end',
+  'bottom-start',
+  'bottom-end',
+  'left-start',
+  'left-end',
+] as const